Add context-guide how to change settings for static front-page
Many users did try to set the custom sidebars for static frontpage and
post-index via the sidebars metabox in the page editor. This caused some
issues/collisions with Sidebar-Location settings ("Archive: Front-Page")
So we hide the input fields in the metabox for the static frontage and
post-index and display some context help instead.